requirements.txt
pandas
polars
plotly
requests
nbformat
matplotlib
python-dotenv
altair